e2e: Extract channel related functions to util/channel.go
We need to create a channel before running tests using subscriptions or
applicationset, and delete it at the end. This code seems like part of
the deployers pacakge. To move it we need to uncouple it from the util
package. The first step is move the related code to a new file.
